Use usize for specifying limits
A bunch of the Alpaca endpoints accept some form of limit parameters.
Right now we specify some of them as u64 and others as usize. None of
these limits is really running risk of exceeding the range of the usize
type on any of the targets that we care about or expect the crate to be
used on. As such, using usize seems like the more general choice and
with this change we switch everything over to using usize.

+++ src/api/v2/orders.rs
@@ -1,4 +1,4 @@
-// Copyright (C) 2019-2021 The apca Developers
+// Copyright (C) 2019-2022 The apca Developers
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 use serde::Serialize;
@@ -34,7 +34,7 @@ pub struct OrdersReq {
   /// The maximum number of orders in response. Defaults to 50 and max
   /// is 500.
   #[serde(rename = "limit")]
-  pub limit: u64,
+  pub limit: usize,
   /// If false the result will not roll up multi-leg orders under the
   /// legs field of the primary order.
   #[serde(rename = "nested")]
